Move AddDeck presenters to new package.
