+++ app/src/main/java/org/mercycorps/translationcards/activity/addDeck/presenter/EnterDeckTitlePresenter.java
@@ -1,4 +1,4 @@
-package org.mercycorps.translationcards.presenter;
+package org.mercycorps.translationcards.activity.addDeck.presenter;
 
 import android.app.Activity;
